📝 Update Python Types Intro to include Optional
