+++ docs_src/python_types/tutorial009.py
@@ -1,7 +1,8 @@
-class Person:
-    def __init__(self, name: str):
-        self.name = name
+from typing import Optional
 
 
-def get_person_name(one_person: Person):
-    return one_person.name
+def say_hi(name: Optional[str] = None):
+    if name is not None:
+        print(f"Hey {name}!")
+    else:
+        print("Hello World")
