+++ docs_src/python_types/tutorial010.py
@@ -1,23 +1,7 @@
-from datetime import datetime
-from typing import List, Optional
-
-from pydantic import BaseModel
-
-
-class User(BaseModel):
-    id: int
-    name = "John Doe"
-    signup_ts: Optional[datetime] = None
-    friends: List[int] = []
+class Person:
+    def __init__(self, name: str):
+        self.name = name
 
 
-external_data = {
-    "id": "123",
-    "signup_ts": "2017-06-01 12:22",
-    "friends": [1, "2", b"3"],
-}
-user = User(**external_data)
-print(user)
-# > User id=123 name='John Doe' signup_ts=datetime.datetime(2017, 6, 1, 12, 22) friends=[1, 2, 3]
-print(user.id)
-# > 123
+def get_person_name(one_person: Person):
+    return one_person.name
